hw2/requirements.txt
torch>=1.12.0  
pytorch-lightning>=1.8.0
wandb
hydra-core>=1.2.0
omegaconf>=2.2.0
torchmetrics>=0.10.0 
pycocotools 
Pillow     